change get data from store in function GetNumberOfOrders

+++ NhapHangV2.Utilities/CatalogueUtilities.cs
@@ -121,55 +121,55 @@ namespace NhapHangV2.Utilities
         /// <summary>
         /// Tất cả
         /// </summary>
-        public decimal? AllOrders { get; set; }
+        public int? AllOrders { get; set; }
         /// <summary>
         /// Chưa đặt cọc
         /// </summary>
-        public decimal? UnDeposit { get; set; }
+        public int? UnDeposit { get; set; }
         /// <summary>
         /// Hủy
         /// </summary>
-        public decimal? Cancel { get; set; }
+        public int? Cancel { get; set; }
         /// <summary>
         /// Đã đặt cọc
         /// </summary>
-        public decimal? Deposit { get; set; }
+        public int? Deposit { get; set; }
         /// <summary>
         /// Chờ duyệt đơn
         /// </summary>
-        public decimal? WaitConfirm { get; set; }
+        public int? WaitConfirm { get; set; }
         /// <summary>
         /// Đã duyệt đơn
         /// </summary>
-        public decimal? Comfirmed { get; set; }
+        public int? Comfirmed { get; set; }
         /// <summary>
         /// Đã mua hàng
         /// </summary>
-        public decimal? PurchaseOrder { get; set; }
+        public int? PurchaseOrder { get; set; }
         /// <summary>
         /// Đã về kho TQ
         /// </summary>
-        public decimal? InChinaWarehoue { get; set; }
+        public int? InChinaWarehoue { get; set; }
         /// <summary>
         /// Đã về kho VN
         /// </summary>
-        public decimal? InVietnamWarehoue { get; set; }
+        public int? InVietnamWarehoue { get; set; }
         /// <summary>
         /// Chờ thanh toán
         /// </summary>
-        public decimal? WaitPayment { get; set; }
+        public int? WaitPayment { get; set; }
         /// <summary>
         /// Khách đã thanh toán
         /// </summary>
-        public decimal? Paid { get; set; }
+        public int? Paid { get; set; }
         /// <summary>
         /// Đã hoàn thành
         /// </summary>
-        public decimal? Completed { get; set; }
+        public int? Completed { get; set; }
         /// <summary>
         /// Đã khiếu nại
         /// </summary>
-        public decimal? Complained { get; set; }
+        public int? Complained { get; set; }
 
 
     }
